split to schedule2
